lint elm and split out type alias (issues with tangled templates)
