+++ protoc-gen-elm/testdata/repeated/expected_output/Repeated.elm
@@ -162,7 +162,6 @@ fooEncoder v =
         ]
 
 
-
 type Foo_NestedEnum
     = Foo_EnumValueDefault -- 0
